move availableCountries from context
